requirements.txt
amqp==5.2.0
annotated-types==0.6.0
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
billiard==4.2.0
cbor2==5.5.1
celery==5.3.6
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron_descriptor==1.4.0
cryptography==41.0.5
cssselect2==0.7.0
defusedxml==0.8.0rc2
distlib==0.3.8
Django==4.2.6
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-cors-headers==4.3.0
django-debug-toolbar==4.2.0
django-dotenv==1.4.2
django-filter==23.3
django-ipware==6.0.2
django-redis==5.4.0
django-seed==0.3.1
django-templated-mail==1.1.1
django-timezone-field==6.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.1
drf-nested-routers==0.93.4
drf-yasg==1.21.7
Faker==20.0.3
filelock==3.13.1
freetype-py==2.3.0
geographiclib==2.0
geopy==2.4.1
gunicorn==21.2.0
html5lib==1.1
humanize==4.9.0
idna==3.4
inflection==0.5.1
kombu==5.3.4
lxml==4.9.3
Markdown==3.5.1
mysqlclient==2.2.1
oauthlib==3.2.2
oscrypto==1.3.0
packaging==23.2
Pillow==10.1.0
platformdirs==4.1.0
prometheus-client==0.19.0
prompt-toolkit==3.0.41
psycopg2-binary==2.9.9
pycairo==1.25.1
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
pygame==2.5.2
Pygments==2.17.2
pyHanko==0.21.0
pyhanko-certvalidator==0.26.2
PyJWT==2.8.0
pyOpenSSL==23.3.0
pypdf==3.17.1
pypng==0.20220715.0
python-bidi==0.4.2
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-ipware==2.0.0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
qrcode==7.4.2
redis==5.0.1
reportlab==4.0.7
requests==2.31.0
requests-oauthlib==1.3.1
rlPyCairo==0.3.0
setuptools==69.0.3
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.0
sqlparse==0.4.4
stripe==7.7.0
svglib==1.5.1
tinycss2==1.2.1
toposort==1.10
tornado==6.4
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
uritemplate==4.1.1
uritools==4.0.2
urllib3==2.0.7
vine==5.1.0
virtualenv==20.25.0
wcwidth==0.2.12
webencodings==0.5.1
wheel==0.42.0
xhtml2pdf==0.2.13
